fix(TUP-18014):Rename folder has "ResourceNotFoundException" in error log list
