+++ main/plugins/org.talend.core.repository/src/main/java/org/talend/core/repository/document/IDocumentationService.java
@@ -10,10 +10,12 @@
 // 9 rue Pages 92150 Suresnes, France
 //
 // ============================================================================
-package org.talend.core.model.repository.documentation;
+package org.talend.core.repository.document;
 
 import org.talend.core.IService;
 import org.talend.core.model.properties.Item;
+import org.talend.core.model.repository.ERepositoryObjectType;
+import org.talend.core.repository.document.IGenerateAllDocumentation;
 
 /**
  * tang class global comment. Detailled comment
@@ -23,5 +25,7 @@ public interface IDocumentationService extends IService {
     public boolean saveDocumentNode(Item item) throws Exception;
 
     public boolean createNodeDocumentationItemFromItem(Item item) throws Exception;
+
+    public IGenerateAllDocumentation getDocGeneratorByProcessType(ERepositoryObjectType objType);
 
 }
